Update terminology around subgraphs
Various log messages and internal symbol names and comments referred to
"federated service" rather than the newer and clearer "subgraph".
